requirements.txt
openai
python-dotenv
requests 
beautifulsoup4 
langchain
faiss-cpu
openpyxl
pymupdf
streamlit
pdfplumber
python-docx
PyPDF2

